# docs/swagger.yaml
basePath: /
definitions:
...
    handler.DriverRiderInfoResponse:
        properties:
            rider:
                properties:
                    name:
                        type: string
                    rating:
                        type: number
                type: object
        type: object
...
    handler.ErrorResponse:
        properties:
            error:
                $ref: '#/definitions/handler.ErrorDetail'
        type: object
...
    handler.EstimatesETAResponse:
        properties:
            distance_meters:
                type: integer
            eta_seconds:
                type: integer
            is_estimate:
                type: boolean
        type: object
    handler.EstimatesPriceResponse:
        properties:
            estimates:
                items:
                    $ref: '#/definitions/handler.Estimate'
                type: array
        type: object
...
    handler.HeatmapResponse:
        properties:
            tiles:
                type: string
        type: object
...
    handler.MessageResponse:
        properties:
            message:
                type: string
        type: object
...
    handler.QueueResponse:
        properties:
            queue: {}
        type: object
...
    handler.deviceRegisterRequest:
        properties:
            platform:
                type: string
            token:
                type: string
        required:
            - platform
            - token
        type: object
    handler.feedbackRequest:
        properties:
            message:
                type: string
            ride_id:
                type: string
        required:
            - message
        type: object
...
host: localhost:8080
info:
    contact: {}
    description: 'Go + Gin ride-hailing backend: auth, rides, driver dispatch, geolocation, navigation, and places.'
    title: Ride-Hailing API
    version: 0.1.0
paths:
...
    /api/v1/devices:
        post:
            consumes:
                - application/json
            operationId: postDevices
            parameters:
                - description: Device registration request
                  in: body
                  name: body
                  required: true
                  schema:
                    $ref: '#/definitions/handler.deviceRegisterRequest'
            produces:
                - application/json
            responses:
                "201":
                    description: Created
                    schema:
                        $ref: '#/definitions/handler.MessageResponse'
                "422":
                    description: Validation error
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
            security:
                - BearerAuth: []
            summary: Register a push notification device
            tags:
                - platform
    /api/v1/devices/{token}:
        delete:
            operationId: deleteDevices-ByID
            parameters:
                - description: Device token
                  in: path
                  name: token
                  required: true
                  type: string
            responses:
                "204":
                    description: No content
            security:
                - BearerAuth: []
            summary: Unregister a push notification device
            tags:
                - platform
...
    /api/v1/driver/rides/{id}/notify-arrival:
        post:
            operationId: postDriver-rides-ByID-notify-arrival
            parameters:
                - description: Ride ID
                  in: path
                  name: id
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/handler.MessageResponse'
            security:
                - BearerAuth: []
            summary: Notify the rider that the driver has arrived
            tags:
                - driver
...
    /api/v1/driver/rides/{id}/rider:
        get:
            operationId: getDriver-rides-ByID-rider
            parameters:
                - description: Ride ID
                  in: path
                  name: id
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/handler.DriverRiderInfoResponse'
            security:
                - BearerAuth: []
            summary: Get rider info for a driver's ride
            tags:
                - driver
...
    /api/v1/driver/rides/queue:
        get:
            operationId: getDriver-rides-queue
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/handler.QueueResponse'
            security:
                - BearerAuth: []
            summary: Get the driver's ride queue
            tags:
                - driver
...
    /api/v1/estimates/eta:
        get:
            description: Returns travel seconds and distance for a route between two coordinates, computed from the live routing engine.
            operationId: getEstimates-eta
            parameters:
                - description: Origin latitude
                  in: query
                  name: from_lat
                  required: true
                  type: number
                - description: Origin longitude
                  in: query
                  name: from_lng
                  required: true
                  type: number
                - description: Destination latitude
                  in: query
                  name: to_lat
                  required: true
                  type: number
                - description: Destination longitude
                  in: query
                  name: to_lng
                  required: true
                  type: number
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/handler.EstimatesETAResponse'
                "422":
                    description: Invalid parameters
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
                "500":
                    description: Routing failed
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
            security:
                - BearerAuth: []
            summary: Get a route-based ETA estimate
            tags:
                - estimates
    /api/v1/estimates/price:
        get:
            description: Computes base + distance + time fare with surge for each vehicle type, using the live routing engine.
            operationId: getEstimates-price
            parameters:
                - description: Pickup latitude
                  in: query
                  name: pickup_lat
                  required: true
                  type: number
                - description: Pickup longitude
                  in: query
                  name: pickup_lng
                  required: true
                  type: number
                - description: Dropoff latitude
                  in: query
                  name: dropoff_lat
                  required: true
                  type: number
                - description: Dropoff longitude
                  in: query
                  name: dropoff_lng
                  required: true
                  type: number
                - description: Limit to one type (sedan|suv|luxury); all types when omitted
                  in: query
                  name: vehicle_type
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/handler.EstimatesPriceResponse'
                "422":
                    description: Invalid parameters
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
                "500":
                    description: Failed to compute estimate
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
            security:
                - BearerAuth: []
            summary: Get price estimates by vehicle type
            tags:
                - estimates
    /api/v1/feedback:
        post:
            consumes:
                - application/json
            operationId: postFeedback
            parameters:
                - description: Feedback request
                  in: body
                  name: body
                  required: true
                  schema:
                    $ref: '#/definitions/handler.feedbackRequest'
            produces:
                - application/json
            responses:
                "201":
                    description: Created
                    schema:
                        $ref: '#/definitions/handler.MessageResponse'
                "422":
                    description: Validation error
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
            security:
                - BearerAuth: []
            summary: Submit user feedback
            tags:
                - platform
...
    /api/v1/geo/eta:
        get:
            description: Returns travel seconds and distance for a route between two coordinates, computed from the live routing engine.
            operationId: getGeo-eta
            parameters:
                - description: Origin latitude
                  in: query
                  name: from_lat
                  required: true
                  type: number
                - description: Origin longitude
                  in: query
                  name: from_lng
                  required: true
                  type: number
                - description: Destination latitude
                  in: query
                  name: to_lat
                  required: true
                  type: number
                - description: Destination longitude
                  in: query
                  name: to_lng
                  required: true
                  type: number
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/handler.EstimatesETAResponse'
                "422":
                    description: Invalid parameters
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
                "500":
                    description: Routing failed
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
            security:
                - BearerAuth: []
            summary: Get a route-based ETA estimate
            tags:
                - estimates
...
    /api/v1/heatmap:
        get:
            operationId: getHeatmap
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/handler.HeatmapResponse'
            security:
                - BearerAuth: []
            summary: Get a heatmap tile (stub)
            tags:
                - platform
    /api/v1/navigation/route:
        get:
            operationId: getNavigation-route
            parameters:
                - description: Origin latitude
                  in: query
                  name: from_lat
                  required: true
                  type: number
                - description: Origin longitude
                  in: query
                  name: from_lng
                  required: true
                  type: number
                - description: Destination latitude
                  in: query
                  name: to_lat
                  required: true
                  type: number
                - description: Destination longitude
                  in: query
                  name: to_lng
                  required: true
                  type: number
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/handler.NavigationRouteResponse'
                "500":
                    description: Routing failed
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
            security:
                - BearerAuth: []
            summary: Get a route between two coordinates
            tags:
                - navigation
    /api/v1/places/autocomplete:
        get:
            operationId: getPlaces-autocomplete
            parameters:
                - description: Latitude
                  in: query
                  name: lat
                  required: true
                  type: number
                - description: Longitude
                  in: query
                  name: lng
                  required: true
                  type: number
                - description: Free-text search query
                  in: query
                  name: q
                  type: string
                - default: 1000
                  description: Search radius in meters
                  in: query
                  name: radius
                  type: number
                - default: 10
                  description: Maximum results (max 50)
                  in: query
                  name: limit
                  type: integer
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/handler.PlacesResponse'
                "422":
                    description: Invalid parameters
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
                "500":
                    description: Query failed
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
            security:
                - BearerAuth: []
            summary: Autocomplete places near coordinates
            tags:
                - places
    /api/v1/places/details:
        get:
            operationId: getPlaces-details
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/handler.PlacesResponse'
            security:
                - BearerAuth: []
            summary: Get place details (stub)
            tags:
                - places
    /api/v1/places/geocode:
        get:
            description: Returns the nearest known place to the given coordinates, used to turn a map pin into an address.
            operationId: getPlaces-geocode
            parameters:
                - description: Latitude
                  in: query
                  name: lat
                  required: true
                  type: number
                - description: Longitude
                  in: query
                  name: lng
                  required: true
                  type: number
                - default: 500
                  description: Search radius in meters (default 500, capped at the configured max)
                  in: query
                  name: radius
                  type: number
            produces:
                - application/json
            responses:
                "200":
                    description: place is null when nothing is found within the radius
                    schema:
                        $ref: '#/definitions/handler.GeocodeResponse'
                "422":
                    description: Invalid parameters
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
                "500":
                    description: Query failed
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
            security:
                - BearerAuth: []
            summary: Reverse-geocode a coordinate into a place
            tags:
                - places
    /api/v1/promotions:
        get:
            operationId: getPromotions
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/handler.PromotionsResponse'
            security:
                - BearerAuth: []
            summary: List available promotions
            tags:
                - promotions
    /api/v1/promotions/apply:
        post:
            operationId: postPromotions-apply
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/handler.MessageResponse'
            security:
                - BearerAuth: []
            summary: Apply a promotion (stub)
            tags:
                - promotions
...
    /api/v1/rider/me/preferences:
        get:
            description: Placeholder for endpoints pending payment integration and other future work.
            operationId: getRider-me-preferences
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/handler.StubResponse'
            security:
                - BearerAuth: []
            summary: Not yet implemented (stub)
            tags:
                - platform
        put:
            operationId: putRider-me-preferences
            responses:
                "200":
                    description: OK
            security:
                - BearerAuth: []
            summary: PUT /api/v1/rider/me/preferences
            tags:
                - rider
...
securityDefinitions:
    BearerAuth:
        description: 'JWT access token. Include as `Authorization: Bearer <token>`.'
        in: header
        name: Authorization
        type: apiKey
swagger: "2.0"